Migrate personnel-form.js to es6
