+++ aidants_connect_common/utils/urls.py
@@ -4,10 +4,10 @@ from django.conf import settings
 def join_url_parts(base: str, *args):
     if len(args) == 0:
         return base
-    rest, tail = args[:-1], args[-1]
+    head, tail = args[:-1], args[-1]
     parts = "/".join(
         [
-            *[arg.removeprefix("/").removesuffix("/") for arg in rest],
+            *[arg.removeprefix("/").removesuffix("/") for arg in head],
             tail.removeprefix("/"),
         ]
     )
