+++ aidants_connect_habilitation/static/js/personnel-form.js
@@ -1,0 +1,60 @@
+"use strict";
+
+import {BaseController} from "./base-controller.js"
+
+(function () {
+    class PersonnelForm extends BaseController {
+        connect() {
+            this.formCountValue = this.managmentFormCountTarget.value;
+            this.formMaxCountValue = this.managmentFormMaxCountTarget.value;
+            const elt = document.querySelector("#issuer-data");
+            this.issuerDataValue = elt ? JSON.parse(elt.textContent) : {};
+            this.controllerInitializedValue = true;
+        }
+
+        formCountValueChanged() {
+            // A bug in Stimulus makes change events to get fired before the controller's `initialize` method is called
+            // which makes the initial form value to get erased. We use the controllerInitializedValue as a flag that
+            // the value has not been initialized yet, here.
+            if (this.controllerInitializedValue) {
+                this.managmentFormCountTarget.value = this.formCountValue;
+                this.mutateVisibility(
+                    this.formCountValue < this.formMaxCountValue, this.addAidantButtonContainerTarget
+                );
+            }
+        }
+
+        onAddAidantButtonClicked() {
+            const template = this.aidantFormTemplateTarget.innerHTML.replace(/__prefix__/gm, `${this.formCountValue}`);
+            this.aidantFormsetTarget.insertAdjacentHTML("beforeend", template);
+            this.formCountValue++;
+        }
+
+        onManagerItsMeBtnClicked() {
+            Object.keys(this.issuerDataValue).forEach(key => {
+                this.managerSubformTarget.querySelector(`[name$='${key}']`).value = this.issuerDataValue[key];
+            });
+        }
+
+        static targets = [
+            "managmentFormCount",
+            "managmentFormMaxCount",
+            "addAidantButtonContainer",
+            "aidantFormTemplate",
+            "aidantFormset",
+            "managerSubform",
+        ]
+
+        static values = {
+            formCount: Number,
+            formMaxCount: Number,
+            controllerInitialized: Boolean,
+            issuerData: Object
+        }
+    }
+
+    new Promise(resolve => window.addEventListener("load", resolve)).then(() => {
+        const application = Stimulus.Application.start();
+        application.register("personnel-form", PersonnelForm);
+    });
+})();
