+++ aidants_connect_habilitation/templates/add_aidants_request.html
@@ -62,6 +62,5 @@
 
 {% block extrajs %}
   {% stimulusjs %}
-  <script src="{% static 'js/utils.js' %}"></script>
-  <script src="{% static 'js/personnel_form.js' %}"></script>
+  <script type="module" src="{% static 'js/personnel-form.js' %}"></script>
 {% endblock %}
