+++ aidants_connect_habilitation/templates/personnel_form.html
@@ -178,8 +178,7 @@
 {% block extrajs %}
   {{ issuer_data|json_script:"issuer-data" }}
   {% stimulusjs %}
-  <script src="{% static 'js/utils.js' %}"></script>
-  <script src="{% static 'js/personnel_form.js' %}"></script>
+  <script type="module" src="{% static 'js/personnel-form.js' %}"></script>
 {% endblock %}
 
 {% block extracss %}
